Feat: add validations

+++ src/socios/socio.service.ts
@@ -1,7 +1,9 @@
+/* eslint-disable prettier/prettier */
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { SocioEntity } from './socio.entity';
+import { BusinessError, BusinessException } from '../shared/errors/errors';
 
 @Injectable()
 export class SocioService {
@@ -15,25 +17,36 @@ export class SocioService {
 	}
 
 	async findOne(id: number): Promise<SocioEntity> {
-		return this.socioRepository.findOne({ where: { id } });
+		const socio = await this.socioRepository.findOne({ where: { id } });
+		if (!socio) {
+			throw new BusinessException('El socio con el id proporcionado no existe', BusinessError.NOT_FOUND);
+		}
+		return socio;
 	}
 
 	async create(socio: SocioEntity): Promise<SocioEntity> {
 		if (!socio.email.includes('@')) {
-			throw new Error('Email inválido');
+			throw new BusinessException('Email inválido', BusinessError.BAD_REQUEST);
 		}
-		return this.socioRepository.save(socio);
+		return await this.socioRepository.save(socio);
 	}
 
 	async update(id: number, socio: SocioEntity): Promise<SocioEntity> {
 		if (!socio.email.includes('@')) {
-			throw new Error('Email inválido');
+			throw new BusinessException('Email inválido', BusinessError.BAD_REQUEST);
 		}
-		await this.socioRepository.update(id, socio);
-		return this.socioRepository.findOne({ where: { id } });
+		const socioPersistido: SocioEntity = await this.socioRepository.findOne({where: {id}});
+		if (!socioPersistido) {
+			throw new BusinessException('El socio con el id proporcionado no existe', BusinessError.NOT_FOUND);
+		}
+		return await this.socioRepository.save({...socioPersistido, ...socio});
 	}
 
 	async delete(id: number): Promise<void> {
+		const socio = await this.socioRepository.findOne({ where: { id } });
+		if (!socio) {
+			throw new BusinessException('El socio con el id proporcionado no existe', BusinessError.NOT_FOUND);
+		}
 		await this.socioRepository.delete(id);
 	}
 }
